Updates source hash for mpfr v4.0.2 and adds missing autoconf macros.
Additional autoconf macros are needed for v4.0.2 which are provided by the
autoconf-archive package. This commit adds a missing spack package for
autoconf-archive and fixes an outdated source hash for mpfr v4.0.2.

It might be better to simply include the autoconf-archive dependency at the
autoconf package level if other packages also need the extension macros.