Excludes SIMD optimization arguments when using PGI toolchain.
The FFTW package attempts to apply SIMD optimization flags when building FFTW.
The flags selected are not controlled by variants to the fftw package spec.
However, when using PGI on ppc64le, unsupported flags are added to the
configuration.

This commit removes targetted optimization options when using a PGI toolchain.
However, it is likely this is only necessary on certain architectures (ppc64le?)
or OSes (RHEL7?) and the logic should be improved before pushing this change to
upstream spack.