Patches broken pre-built docs out of build
The RDMA-Core releases ship with pre-built documentation in case pandoc
and rst2man are not available on the system. It is not possible to
disable documentation at configuration time and the build will fail if
documentation cannot be built or installed. However, the pre-build
documentation is not complete (at least for infiniband-diags) and
rdma-core's CMake install process will attempt to copy pre-built
documentation that does not exist at install time, causing a fatal
error.

Adding a dependency on pandoc makes the DAG for this package extremely
large (pulling in many X11 dependencies). This is bound to cause
problems for non-GCC toolchains. So rather than have the spack package
build the docs from source, we will rely on the pre-built documentation
but exclude the attempted installation of the missing infiniband-diags
man pages.