Adds missing PIC flag for MPI builds with PGI.
PGI doesn't automatically propagate the PIC flag when building with MPI
support and autoconf does not detect the appropriate PIC flag when using
the mpif90 spack wrapper.

This commit forces FCFLAGS (respected by the mpif90 wrapper) to contain
the appropriate PIC flag when building with that option and MPI support
for all compilers.