Updates NetCDF-Fortran version; adds GCC10 build error fix.
GCC10 fails to build netcdf-fortran and parallel-netcdf due to a
issue common to both packages that is known to the upstream developers
(at least for netcdf-fortran):

https://github.com/Unidata/netcdf-fortran/issues/212

NetCDF-fortran has fixed the issue in a new release as of early June 2020
(https://github.com/Unidata/netcdf-fortran/pull/250) but the problem has
not yet been patched in parallel-netcdf.

The latest version of netcdf-fortran (v4.5.3) requires netcdf-c >=
v4.7.4 so the base netcdf-c package is also updated to the latest
release.