Patches netlib-scalapack to solve confusion between PGI and Intel.
The netlib-scalapack CMakeLists have a faulty check to identify if the
Intel compiler suite is being used. It checks if the substring 'ifort'
matches the fortran compiler binary. Unfortunately, this also catches
the PGI fortran compiler 'pgifort'. This causes the build system to
inject Intel-specific compiler options into the build. Since
unrecognized arguments are generally fatal to PGI, this causes problems
when trying to build the package with PGI.

Instead, the developers probably want to check the
`CMAKE_Fortran_COMPILER_ID` for `Intel` which is unique for the Intel
toolchain. This commit adds a patch for this to spack builds. This patch
should probably be generalized better for future versions and applied as
a patch function. Or, better yet, contributed to the upstream Scalapack
codebase.