LSF systems must use own PMIx implementation.
IBM's PMIx libraries are not provided with LSF and instead are
incorporated in CSM or SMPI. OMPI cannot reliably find the necessary
external PMIx implementation when on certain systems using LSF.

Therefore, we force the use of internal PMIx when using LSF.

However, the LSF libdirs options are also not reliably found due to the
way IBM packages LSF headers and libraries in separate locations. The
second block of code in this commit is experimental and doesn't
necessarily work.