Adds ability to set module properties via Spack
Module properties for use with LMOD (described in full at
https://lmod.readthedocs.io/en/latest/145_properties.html) can now be added to
modulefiles by customizing the spack module context on a per-spec basis.
Arbitrary property keys can added in the following format (using a spack
environment as an example):

```
spack:
  modules:
    lmod:
      <package>:
        module_property_tags:
            <spec1>:
                <property_key1>: <property_value>
                <property_key2>: <property_value>
            <spec2>:
                <property_key>: <property_value>
```

However, each key applied to module can only have one value in LMOD. So if
multiple specs in `module_property_tags` apply to a build and each provide a
different value for a given key, the first one inserted (indeterministically)
into the module will be used by LMOD. Duplicate key-value properties are
filtered out before writing the modulefile. Only lua/LMOD modules will actually
have property statements written to the modulefile.

And, for properties to be used by LMOD, the `lmodrc` file used on the system
must be configured to read and correctly interpret any properties not in the
default set.