Updates remaining ROCm packages to support v4.2.0
miopen-hip, rocfft (and it's dependent hipfft), and rocm-openmp-extras
needed either ROCm v4.2.0 versions added or special patches to prevent
them from searching for dependnecies in `/opt/rocm`. Searching
`/opt/rocm` fails if only older versions like v4.1.0 are installed on
the system through RPM distributed packages. Which indicates that the
previous installations of these packages are likely broken in that they
are not independent from system-installed ROCm. If that's the case, they
will likely malfunction if and when ROCm installed through RPMs is added
to a system.