Reverts llvm-amdgpu dependency OMP option
Clang-12 refuses to build if both the core openmp and rocm-openmp-extras
project targets are enabled since they both provide the same openmp
target. So it is necesary for this package to depend on
llvm-amdgpu~openmp for ROCm 4.2.0+.

It is really necessary for this package to be merged with llvm-amdgpu in
order to prevent duplicate builds of other ROCm components with mixed
dependency on llvm-amdgpu variants. Also, the code provided by this
package really is an extension to llvm-amdgpu and shouldn't exist in a
separate prefix.