Fixes broken symlink in ROCm-smi-lib.
ROCm-SMI-lib depends on an rsmiBindings python module that, when
the package is installed in a separate prefix from the rest of the ROCm
stack, is available to the rocm-smi executable (script) via a symlink
that points to the wrong location.

This commit fixes the symlink to point to where this python module
actually exists in the package's install prefix.