Fixes rocprof use of incorrect relative path for ROCM_PATH.
The CLI shell scripts for rocprof look for ROCm components relative to
the location of the `rocprof` and/or `bin/rpl_run.sh` scripts. When this
package is installed in an arbitrary prefix separate from the rest of
the ROCm stack (unlike a "normal" non-spack installation of ROCm in
`/opt/rocm`), these necessary components cannot be found.

This commit changes the rocprof CLI interface script to use ROCM_PATH
environment variables to locate it's dependencies. This works when
rocprofiler-dev is exposed to users through a "spack view" along with
all other ROCm component dependencies as if ROCm was installed by spack
into a single common prefix. However, it will not work if such a view
does not exist. In the event such a view does not exist, it will
fallback to looking in `/opt/rocm` for it's dependencies. This will not
work if ROCm (of matching version) is not installed there.

Ideally, instead of using ROOT_DIR to locate dependency components, each
dependency path would be hard-coded at install-time to point to where
the dependency actually exists. This is possible for the HSA runtime
libs, but due to a circular dependency on roctracer-dev, we cannot do
this for the roctracer libs.

The issues of
- inconsistancy in locating dependency paths w.r.t. other ROCm
  components (searching relative to this package's location and not
  relying on ROCM_PATH or other standard envvars)
- the circular dependency on roctracer (which is only broken by having
  ROCm already installed on the system at `/opt/rocm`)
- the lack of support in this package's CMakeLists to specify alternate
  locations for dependencies not in `/opt/rocm` and correctly hard-code
  their locatins in this package's shell scripts
- complete independence of the ROCm stack on `/opt/rocm` and being
  co-installed in a common prefix.

must be addressed in all the packages in the ROCm stack as a whole OR
ROCm should be packaged as a stand-alone monolithic package or as a
monolithic core that is extensible with individual packages for
components like ROCm runtime math libraries (rocblas/hipblas,
rocfft/hipfft, etc.).