Patches HPCToolkit build-time python script
On systems without a generic `python` executable in the PATH,
HPCToolkit's 'stringify.py' script needed at build-time fails to
execute. Since python2 is deprecated and all systems moving forward
should have a python3 interpreter in their sysroot, we patch this script
(which appears to use python3 syntax) to use python3 explicitly.

A better solution in the long-term would be to explicitly add python3 as
a build-time dependency to the spack package.