Fixes incorrect CMake compiler ID used by Scalapack
Netlib-scalapack's CMakeLists.txt uses regex matching on the compiler
name to ID the compiler rather than using CMake's builtin and
unabmiguous compiler ID macros. This commit adds a patch to correct this
so that, for example, the OMPI fortran wrapper `mpifort` using a
gfortran, pgfortran, or xlf backend compiler isn't identified as the
intel fortran compiler. Most toolchain fortran drivers ignore the
intel-specific commandline arguments added when scalapack thinks Intel
ifort is being used but this is a fatal error when PGI is the MPI
fortran wrapper backend.