Drops `kokkos@develop` version from HIP-enabled vtk-m builds
Kokkos v3.4.00, which is now available, adds hip/rocm support and it is
no longer necessary for vtk-m to depend on an arbitrary development
version of kokkos when hip/rocm support is used. Rather than depend
on kokkos@develop, the package is updated to depend on tagged kokkos
release versions that support ROCm.