Fixes patch for install error when AML documentation is not built
The AML spackage has an undeclared dependency on doxygen needed to build
the docs. However, if and when doxygen is not available, an attempt is
made to install pre-built docs shipped with the source without first
creating the appropriate destination dir for them under the install
prefix causing an install error for an otherwise successful build. This
issue affects at least AML version 0.1.0.

This commit fixes this problem by patching the Makefile template to
create the pre-built install directory before attempting to copy the
files shipped with the source.