Fixes OLCF patch for HPCToolkit stringify util for new package version
The HPCToolkit previously used a python script
`src/tool/hpcstruct/stringify.py` at configure-time which did not work
on systems where `python` in the environment was a python2 interpretor.
The OLCF added a patch to correct this behavior on some systems with
multiple python interpretors in the PATH to follow PEP-394 and applied
the patch indiscriminately.

Subsequent releases of HPCToolkit have replaced the python script with a
shell script and it is now necessary to check if the target python
script exists before applying the patch.