Fixes fortran misconfiguration in Conduit spackage
The conduit spackage is modified to produce a correct configuration file
when attempting to enable fotran support or raise an error if a suitable
fortran compiler cannot be determined.

Although `$SPACK_FC` is set in the build environment when building with
fortran support, the conduit spackage checks to see if the value is an
actual file. In the case where the value of the envar is simply the
fortran compiler binary name that would be found in the `$PATH`and is
not the full path to the binary, this check fails and the conduit
spackage configures the build without fortran support enabled. Our patch
first gets the full path to the value of `$SPACK_FC` by searching the
`$PATH` (which works both when only a binary name is set or when a full
path to the binary is set) and then testing if the binary actually
exists as a file. If a valid fortran compiler is not found on disk, an
exception is raised to prevent a misconfigured install that otherwise
appears successful.