Adjusts Cray platform for ORNL EA-systems
The Cray Programming Environment since cpe v21.07 loads cray-mpich along
with any PrgEnv-* meta modules. Since we want to control the version of
cray-mpich that is loaded using a `cray-mpich` external-only spack
package, we must unload cray-mpich along with other PrgEnv-*
dependencies when setting up the build environment.

cray-libsci at v21.08.1.2 also has a hard dependency on cray-mpich so
when declaring external modules for `cray-mpich` and `cray-libsci`
spackages, it is necessary to load a specific cray-mpich module in each.
It is also necessary to ensure the cray-mpich module matches when both
dependencies are needed in a build.

This commit also adds an architecture alias map for additional Cray
compiler wrapper targets.