Adds support for Zen3 CPU targets
Kokkos and trilinos (when it depends on kokkos) must be aware of AMD
zen3 CPU targets when the spack arch targets that microarchitecture.

For the time being, building on zen3 configures kokkos to optimize for
the zen2 ISA subset of instructions until Kokkos is released with
specific optimizations for Zen3 CPUs.