Disables Gingko smoke tests when ROCm support is enabled
The smoke tests attempt to build without the correct environment for
linking against ROCm components when ROCm is enabled and fail for what
is otherwise a successful build. This commit disables the smoke tests
when ROCm support is enabled until the test build environment can be
corrected in a more appropriate way in the upstream ginkgo spackage.