Improves Cray backend target detection performance
If the backend target is specified via an environment variable, that
value is used immediately. However, when the default backend target is
not explictly set with an environment variable, it is detected through
default loaded target modules, spack archspec, and finally the
python-detected machine platform.

Originally, this detection sequence was executed every time the Cray
platform class was instantiated since the _default_target_from_env()
routine was called as the default value of os.environ.get(). Detecting
the default value from the environment is an expensive operation which
would be invoked one or more times per package to be installed when
running on a Cray.

This commit caches the detected backend target value as a singleton and
only invokes the detection method once if it is ever needed when
concretizing or installing specs in an environment on a Cray.