Adds Frontier's GPU arch to ROCm targets and select packages.
The gfx90a GPU target is added to the ROCm package and to VTK-m's valid
targets. Kokkos still needs to be made gfx90a-aware but this requires a
newer version (3.5.0) of Kokkos than is part of the current E4S release.

Adding the newer version will technically diverge the E4S packages
installed from the releases' prescribed configurations.