Fixes Boost bootstrapping with Cray PrgEnv-amd
Cray Shasta systems having AMD GPUs offer a Cray PrgEnv-amd compiler
environment that uses amdclang behind the normal cray compiler wrappers.

Spack does not currently have a compiler definition for amdclang.
Adding a Cray PrgEnv-amd compiler spec to Spack as vanilla '%clang' is
sufficient in most cases. However, Boost's bootstrapping process with a
"clang" toolset expects to find a binary in the path named "clang++"
capable of supporting C++11. PrgEnv-amd, despite being an LLVM/clang
toolchain with a frontend compatible with vanilla clang, is exposed in
the environment as "amdclang++" or "CC". In order to bootstrap Boost in
this environment, we use the available OS-provided GCC in the same
manner as building Boost with armclang or fj toolchains.