Fixes incomplete type comparison bug causing excessive env file re-writes.
When comparing raw and new environment file yaml contents,
`spack.environment.Environment._update_and_write_manifest` attempts to
compare two yaml objects for equivalence using
`spack.environment.yaml_equivalent`. This recursively checks two yaml
documents element-wise but only expects values that are either dicts,
lists, or strings. If the value under comparison is not a dict or list,
it is assumed to be string. However, other types may exist in the
documents (including `bool`, `int`, and `float` type values).
When such a type is encountered, the two yaml documents are flagged as
being different and the new yaml document is written to disk.

Due to a separate, as-of-yet unidentified, bug in either
`spack.utils.spack_yaml` or `ruayaml.yaml`, writing the environment
`spack_yaml.dump_config` will necessarily add a newline before every
non-trailing comment in a map or list section of the document at least
whenever the comment is indented to match its surroundings. This causes
new whitespace added everytime the environment is partially or fully
concretized and makes preserving an clean and readable authoritative
environment file in version control difficult, especially when used in
CI jobs where the tracked files are not expected to be changed relative
to a MR commit.

The following fix catches only the cases when elements in the yaml
document are truly different in both type and value and significantly
reduces the number of times a spack environment file is unnecessarily
over-written and therefore also reduces (but does not eliminate) the
amount of unwanted whitespace added around comments in VCS-tracked
environment files.