Updates adds missing autoconf macros to mpfr build-time and runtime environments
Additional autoconf macros are needed for v4.0.2 which are provided by
the autoconf-archive package. This commit ensures the autoconf-archive
macros are available at buildtime for mpfr dependents and at mpfr
runtime.

It might be better to simply include the autoconf-archive dependency at
the autoconf package level if other packages also need the extension
macros.