Adds patch for nvhpc toolchain to successfully build OMPI
See https://github.com/open-mpi/ompi/issues/9444 for details. The patch
changes an m4 file so running a reconf stage is necessary even for the
distribution releases.

Autoreconf of ompi up to v4.1.1 *requires* automake v1.15.x,
unfortunately. So the default v1.16.x available in Spack must be
over-ridden in the spec and possibly dependency builds if all
OMPI-dependent builds done with a given nvhpc toolchain will share a
single build of a given OMPI release.

It is also necessary to build some components with `-fPIC` but it was
not clear which toolchain language drivers were missing the flag (which
is enabled by default). The PIC flag modifications in this commit are
probably excessive but seem to work.