Patches HPCToolkit build-time python script
On systems without a generic `python` executable in the PATH,
HPCToolkit's 'stringify.py' script needed at build-time for older
versions fails to execute. Since python2 is deprecated and all systems
moving forward should have a python3 interpreter in their sysroot, we
patch this script (which appears to use python3 syntax) to use python3
explicitly when it exists and is necessary.