Allow libs to be found in lib64 subdir.
When using external OpenSSL, the libs are not always in `$prefix/lib` in
the external build. Sometimes, the libraries exist in `$prefix/lib64`.
Packages which must declare the location of the SSL libraries outside of
using pkgconfig need the `openssl.libs` method to return valid paths.

Even though recursive search for the necessary libs should search both
paths, it seems there are cases where, when using external OpenSSL, this
search is incomplete.