FIXME: Ignores all errors when building on Crays with GCC11+
For some reason, when building GraphBLAS using GCC11.2.0 on a Cray with
CPE21.08, `-Werror=misleading-indentation` is in effect and all
combinations of arguments to selectively ignore misleading indentation
warnings/errors are not respected. Passing `-w` is a heavy hammer but it
seems the CPE provided GCC11.2.0 toolchain cannot ignore just misleading
indentations.

This commit should be removed if and when a future release of CPE with
fixed GCC11.2.0 is available.