FIXME: Reduces max processes used during environment concretization
When using the clingo solver, spack will run clingo with a pool of
process threads equal in number to the smallest value of either:

- the number of specs to solve,
- the available CPUs, or
- a hard-coded maximum limit of 16

threads. On some shared resources having more than 16 CPUs and specs to
solve, the hard-coded maximum thread limit of 16 can consume more memory
than is available under per-user control group memory limitations
policies enforced by system admins. Ideally, the maximum number of
threads used by clingo would be runtime-configurable, perhaps following
the `build_jobs` setting that already exists to limit parallelism in
package builds.

See upstream issue https://github.com/spack/spack/issues/29464 for a
proposed feature request to make the maximum number of clingo threads
configurable. Until such a feature is implemented upstream, it is
necessary to hard-code a lower value on the maximum number of processes.

This commit reduces the maximum number of threads used when solving an
environment to 10. However, this commit should be modified to use the
lesser of either `spack.config.build_jobs` or the original default of 16
instead of hard-coding a new lower default value. Or this commit should
be dropped when a solution for issue 29464 is implemented upstream.