Explicitly disables SVG support in Cairo when not requested
The Cairo build system automatically enables SVG support at
configure-time if detects the OS can support it. However, on RHEL8
systems, the location librsvg headers are in version-suffixed subdirs of
the normal header search path such that at build-time, the headers
cannot be located without special modifications to the header search
path. We don't generally want to use the OS-provided librsvg
implementation so unless the SVG variant is requests, this commit
explicitly disables SVG support.

Additionally, librsvg built by spack depends on cairo in a circular
dependency loop so it is not actually possible to build cairo with SVG
support entirely with spack at this time.