Bypasses broken ICL cmake module for detecting BLAS
The ICL cmake module shipped with lapack++ for detecting the BLAS
implementation to use is broken. Instead of using that macro, the
lapack++ spackage has been modified to use the fallback CMake-builtin
BLAS detection module and explicitly pass the BLAS libraries offered by
the spack spec.

This commit also adds the flag for enabling trailing underscore name
mangling when building with gfortran.