FIXME: Disables broken smoke tests
Smoke tests for Sundials and tasmanian fail post-install due to missing
test files in the final install prefix. These problems are not a concern
for some installations and shouldn't be fatal to spack at install time
at the user's discrection. For now, these smoke tests are completely
disabled to prevent interruptions to software deployment at the OLCF.