Fixes circular dependency in librsvg and cairo+svg
There is a dependency loop with `cairo+svg` and `librsvg`. `librsvg`
[necessarily depends on cairo](https://gitlab.gnome.org/GNOME/librsvg/-/blob/main/COMPILING.md#installing-dependencies-for-building).
And, though the librsvg docs state it needs cairo built with PNG support
(not currently declared in the spack package), it is apparent when
building librsvg that it also needs cairo configured with the flag
`--enable-svg=yes` in order to have svg-related parts of the cairo
interface used by librsvg. That is, `librsvg ^cairo+svg` (also not in
the current spack package for `librsvg`).

However, since the spack package for `cairo+svg+png` depends on librsvg,
having librsvg depend on `cairo+png+svg` introduces a dependency loop.

This affects builds of `gtkplus` and its dependents. In the past, this
has been side-stepped by using external OS-provided GTK but this
approach does not work for spack users interested in not relying on
external GTK.

It is not clear if this commit is the best way to break this dependency
loop.  For what it's worth, this commit builds these pacakges in a way
similar to how Archlinux packages cairo, librsvg, and gtkplus.

One-off tests of this commit have successfully built `cairo+svg+png`,
`librsvg ^cairo+svg+png` and a dependent `gtkplus` used to build
`imagemagick`, `emacs+X`, and other applications. These dependents
appear to run successfully and pass high-level tests.

But it has not been checked if via careful study of the build logs to be
sure cairo isn't picking up svg-related libs from the OS and that this
is truly the correct way to break the circular dependency.  The cairo
and librsvg documentation isn't very clear about the dependency chain.

See the comment chain in [spack PR 14092](https://github.com/spack/spack/pull/14092)
for addition information.