Disables error-on-warning when building libzmq with GCC12+
G++ 12.1.0 is more sensitive to buffer overrun errors than previous
versions. Libzmq 4.3.4 includes a unit test helper function that loops
over an array and checks if the last+1 element is valid by means of a
pointer comparison that g++ 12.1.0 catches and warns about.  Therefore,
it is necessary to not fail the build when this warning is raised by
disabling `-Werror` when building libzmq up to v4.3.4 with GCC 12.